Update BuildTIme component
