+++ components/models/elements/BuildTime.tsx
@@ -1,44 +1,19 @@
-import { SFC, Fragment } from 'react';
+import { FunctionComponent } from 'react';
 import { useFetch } from 'react-async';
+import { makeTimeFriendly } from '../../../utils/functions/convertTime';
 
 interface iBuildTime {
   id: string;
-}
-
-function convertTime(time: String): string {
-  // verify we even have something first
-  if (!time) {
-    return `N/A`;
-  }
-
-  // step 1: remove the PT/FT
-  const stripped = time.slice(2);
-
-  // 0S only shows if there is no time logged but a project has been created, so just return none
-  if (stripped == `0S`) {
-    return `None`;
-  }
-
-  const regex = new RegExp(/((\d+)H)?(\d+)M(\d+)S/);
-  const splitString = stripped.match(regex);
-  if (splitString) {
-    const hours = splitString[2];
-    const minutes = splitString[3];
-
-    return `${hours || 0} Hours ${minutes} Minutes`;
-  }
-
-  return `N/A`;
 }
 
 interface iFetchData {
   duration: string;
 }
 
-const BuildTime: SFC<iBuildTime> = ({ id }) => {
+const BuildTime: FunctionComponent<iBuildTime> = ({ id }) => {
   // immediately return if id is null
   if (id === null) {
-    return <Fragment>N/A</Fragment>
+    return null;
   }
 
   const headers = {
@@ -53,14 +28,12 @@ const BuildTime: SFC<iBuildTime> = ({ id }) => {
     }
   );
 
-
-  if (isLoading) return <Fragment>---</Fragment>;
   if (error) {
-    console.log(error.message);
-    return <Fragment>N/A</Fragment>;
+    console.log(`Unable to fetch from clockify: ${error.message}`);
+    return null;
   }
 
-  if (data) return <Fragment>{convertTime(data.duration)}</Fragment>;
+  if (data) return <span>Build Time: {isLoading ? '-----' : makeTimeFriendly(data.duration)}</span>;
 
   return null;
 };
