+++ tsconfig.json
@@ -23,7 +23,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "declaration": true,
-    "outDir": "dist",
+    "outDir": "dist"
   },
   "exclude": [
     "node_modules"
